fix: lint files
